+++ sphinx-docs/source/changelog.rst
@@ -1,19 +1,22 @@
 Change Log
 ==========
 
-v2.2 (TBD)
+v2.2 (2021-03-01)
 --------------------
 
 *New Features*
 
 - Can read molecules from GROAMCS topology
 - Can compute one-hot encoding for different bead type interactions in CG systems
+- Can generate mapped exclusion lists in CG systems
 
 *Enhancements*
 
 - Added jupyter notebook testing to continuous integration
+- `iter_from_trajectory` can run in a user-defined time range
+- Revised to work for TF 2.4
+- Can add exclusion lists when computing nlists using htf
 
-*Bug Fixes*
 
 v2.1 (2021-02-04)
 --------------------